# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/payments:
    get:
      operationId: listAdminPayments
      tags: [admin]
      summary: List payments (optionally filtered by status and/or userId)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/PaymentStatus"
        - name: userId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminPayment"
  /admin/payments/{paymentId}/confirm:
    post:
      operationId: confirmPayment
      tags: [admin]
      summary: Confirm a payment and activate the subscription
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Confirmed payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /admin/payments/{paymentId}/reject:
    post:
      operationId: rejectPayment
      tags: [admin]
      summary: Reject a payment
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentReject"
      responses:
        "200":
          description: Rejected payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /admin/payments/{paymentId}/refund:
    post:
      operationId: refundPayment
      tags: [admin]
      summary: Record an external payment refund or chargeback and reverse its referral commission
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentRefund"
      responses:
        "200":
          description: Refunded payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
...
components:
  schemas:
...
    PaymentProvider:
      type: string
      # freekassa is legacy-read-only: integration removed 2026-07-17, kept so
      # historical payment rows still serialize/validate.
      # balance: instant debit from the user's wallet balance (no moderation queue).
      # free_grant: an auditable zero-cost admin grant; never earns referral commission.
      enum: [manual_sbp, yookassa, yoomoney, freekassa, balance, free_grant]
    PaymentStatus:
      type: string
      enum: [pending, confirmed, rejected, refunded]
    PaymentType:
      type: string
      enum: [subscription, extra_device_slot, balance_topup, extra_traffic]
    Payment:
      type: object
      required: [id, subscriptionId, type, provider, amountRub, status, reference, createdAt]
      properties:
        id: { type: integer }
        subscriptionId: { type: ["integer", "null"] }
        type:
          $ref: "#/components/schemas/PaymentType"
        provider:
          $ref: "#/components/schemas/PaymentProvider"
        amountRub: { type: integer }
        # For type === extra_traffic only: how many GB this order grants.
        extraTrafficGb: { type: ["integer", "null"] }
        status:
          $ref: "#/components/schemas/PaymentStatus"
        reference: { type: string }
        userNote: { type: ["string", "null"] }
        # Whether the user has attached a payment confirmation screenshot.
        # The image itself is fetched separately via GET .../screenshot/image.
        hasScreenshot: { type: boolean }
        rejectionReason: { type: ["string", "null"] }
        refundKind: { type: ["string", "null"], enum: [refund, chargeback, null] }
        refundReason: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        confirmedAt: { type: ["string", "null"], format: date-time }
        refundedAt: { type: ["string", "null"], format: date-time }
    AdminPayment:
      type: object
      required:
        [id, subscriptionId, userId, userEmail, planName, type, provider, amountRub, status, reference, createdAt]
      properties:
        id: { type: integer }
        subscriptionId: { type: ["integer", "null"] }
        userId: { type: integer }
        userEmail: { type: string }
        planName: { type: ["string", "null"] }
        type:
          $ref: "#/components/schemas/PaymentType"
        provider:
          $ref: "#/components/schemas/PaymentProvider"
        amountRub: { type: integer }
        extraTrafficGb: { type: ["integer", "null"] }
        status:
          $ref: "#/components/schemas/PaymentStatus"
        reference: { type: string }
        userNote: { type: ["string", "null"] }
        hasScreenshot: { type: boolean }
        rejectionReason: { type: ["string", "null"] }
        refundKind: { type: ["string", "null"], enum: [refund, chargeback, null] }
        refundReason: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        confirmedAt: { type: ["string", "null"], format: date-time }
        refundedAt: { type: ["string", "null"], format: date-time }
...
    PaymentReject:
      type: object
      properties:
        reason: { type: string }
    PaymentRefund:
      type: object
      properties:
        kind:
          type: string
          enum: [refund, chargeback]
          default: refund
        reason: { type: string, maxLength: 500 }
...