# lib/api-zod/src/generated/types/payment.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentProvider } from './paymentProvider';
import type { PaymentRefundKind } from './paymentRefundKind';
import type { PaymentStatus } from './paymentStatus';
import type { PaymentType } from './paymentType';

export interface Payment {
  id: number;
  /** @nullable */
  subscriptionId: number | null;
  type: PaymentType;
  provider: PaymentProvider;
  amountRub: number;
  /** @nullable */
  extraTrafficGb?: number | null;
  status: PaymentStatus;
  reference: string;
  /** @nullable */
  userNote?: string | null;
  hasScreenshot?: boolean;
  /** @nullable */
  rejectionReason?: string | null;
  /** @nullable */
  refundKind?: PaymentRefundKind;
  /** @nullable */
  refundReason?: string | null;
  createdAt: Date;
  /** @nullable */
  confirmedAt?: Date | null;
  /** @nullable */
  refundedAt?: Date | null;
}